static extension : ones that got away
