+++ src/Microsoft.ML.StaticPipe/TransformsStatic.cs
@@ -11,6 +11,7 @@ using Microsoft.ML.Transforms;
 using Microsoft.ML.Transforms.Conversions;
 using Microsoft.ML.Transforms.FeatureSelection;
 using Microsoft.ML.Transforms.Projections;
+using Microsoft.ML.Transforms.Text;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -1363,5 +1364,158 @@ namespace Microsoft.ML.StaticPipe
             }
         }
         #endregion
+    }
+
+    /// <summary>
+    /// Extension methods for the static-pipeline over <see cref="PipelineColumn"/> objects.
+    /// </summary>
+    public static class NAIndicatorStaticExtensions
+    {
+        private interface IColInput
+        {
+            PipelineColumn Input { get; }
+        }
+
+        private sealed class OutScalar<TValue> : Scalar<bool>, IColInput
+        {
+            public PipelineColumn Input { get; }
+
+            public OutScalar(Scalar<TValue> input)
+                : base(Reconciler.Inst, input)
+            {
+                Input = input;
+            }
+        }
+
+        private sealed class OutVectorColumn<TValue> : Vector<bool>, IColInput
+        {
+            public PipelineColumn Input { get; }
+
+            public OutVectorColumn(Vector<TValue> input)
+                : base(Reconciler.Inst, input)
+            {
+                Input = input;
+            }
+        }
+
+        private sealed class OutVarVectorColumn<TValue> : VarVector<bool>, IColInput
+        {
+            public PipelineColumn Input { get; }
+
+            public OutVarVectorColumn(VarVector<TValue> input)
+                : base(Reconciler.Inst, input)
+            {
+                Input = input;
+            }
+        }
+
+        private sealed class Reconciler : EstimatorReconciler
+        {
+            public static Reconciler Inst = new Reconciler();
+
+            private Reconciler() { }
+
+            public override IEstimator<ITransformer> Reconcile(IHostEnvironment env,
+                PipelineColumn[] toOutput,
+                IReadOnlyDictionary<PipelineColumn, string> inputNames,
+                IReadOnlyDictionary<PipelineColumn, string> outputNames,
+                IReadOnlyCollection<string> usedNames)
+            {
+                var columnPairs = new (string input, string output)[toOutput.Length];
+                for (int i = 0; i < toOutput.Length; ++i)
+                {
+                    var col = (IColInput)toOutput[i];
+                    columnPairs[i] = (inputNames[col.Input], outputNames[toOutput[i]]);
+                }
+                return new MissingValueIndicatorEstimator(env, columnPairs);
+            }
+        }
+
+        /// <summary>
+        /// Produces a column of boolean entries indicating whether input column entries were missing.
+        /// </summary>
+        /// <param name="input">The input column.</param>
+        /// <returns>A column indicating whether input column entries were missing.</returns>
+        public static Scalar<bool> IsMissingValue(this Scalar<float> input)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            return new OutScalar<float>(input);
+        }
+
+        /// <summary>
+        /// Produces a column of boolean entries indicating whether input column entries were missing.
+        /// </summary>
+        /// <param name="input">The input column.</param>
+        /// <returns>A column indicating whether input column entries were missing.</returns>
+        public static Scalar<bool> IsMissingValue(this Scalar<double> input)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            return new OutScalar<double>(input);
+        }
+
+        /// <summary>
+        /// Produces a column of boolean entries indicating whether input column entries were missing.
+        /// </summary>
+        /// <param name="input">The input column.</param>
+        /// <returns>A column indicating whether input column entries were missing.</returns>
+        public static Vector<bool> IsMissingValue(this Vector<float> input)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            return new OutVectorColumn<float>(input);
+        }
+
+        /// <summary>
+        /// Produces a column of boolean entries indicating whether input column entries were missing.
+        /// </summary>
+        /// <param name="input">The input column.</param>
+        /// <returns>A column indicating whether input column entries were missing.</returns>
+        public static Vector<bool> IsMissingValue(this Vector<double> input)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            return new OutVectorColumn<double>(input);
+        }
+
+        /// <summary>
+        /// Produces a column of boolean entries indicating whether input column entries were missing.
+        /// </summary>
+        /// <param name="input">The input column.</param>
+        /// <returns>A column indicating whether input column entries were missing.</returns>
+        public static VarVector<bool> IsMissingValue(this VarVector<float> input)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            return new OutVarVectorColumn<float>(input);
+        }
+
+        /// <summary>
+        /// Produces a column of boolean entries indicating whether input column entries were missing.
+        /// </summary>
+        /// <param name="input">The input column.</param>
+        /// <returns>A column indicating whether input column entries were missing.</returns>
+        public static VarVector<bool> IsMissingValue(this VarVector<double> input)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            return new OutVarVectorColumn<double>(input);
+        }
+    }
+
+    /// <summary>
+    /// Extension methods for the static-pipeline over <see cref="PipelineColumn"/> objects.
+    /// </summary>
+    public static class TextFeaturizerStaticExtensions
+    {
+        /// <summary>
+        /// Accept text data and converts it to array which represent combinations of ngram/skip-gram token counts.
+        /// </summary>
+        /// <param name="input">Input data.</param>
+        /// <param name="otherInputs">Additional data.</param>
+        /// <param name="advancedSettings">Delegate which allows you to set transformation settings.</param>
+        /// <returns></returns>
+        public static Vector<float> FeaturizeText(this Scalar<string> input, Scalar<string>[] otherInputs = null, Action<TextFeaturizingEstimator.Settings> advancedSettings = null)
+        {
+            Contracts.CheckValue(input, nameof(input));
+            Contracts.CheckValueOrNull(otherInputs);
+            otherInputs = otherInputs ?? new Scalar<string>[0];
+            return new TextFeaturizingEstimator.OutPipelineColumn(new[] { input }.Concat(otherInputs), advancedSettings);
+        }
     }
 }
