+++ src/Microsoft.ML.Transforms/MissingValueIndicatorTransformer.cs
@@ -477,136 +477,4 @@ namespace Microsoft.ML.Transforms
             return new SchemaShape(result.Values);
         }
     }
-
-    /// <summary>
-    /// Extension methods for the static-pipeline over <see cref="PipelineColumn"/> objects.
-    /// </summary>
-    public static class NAIndicatorExtensions
-    {
-        private interface IColInput
-        {
-            PipelineColumn Input { get; }
-        }
-
-        private sealed class OutScalar<TValue> : Scalar<bool>, IColInput
-        {
-            public PipelineColumn Input { get; }
-
-            public OutScalar(Scalar<TValue> input)
-                : base(Reconciler.Inst, input)
-            {
-                Input = input;
-            }
-        }
-
-        private sealed class OutVectorColumn<TValue> : Vector<bool>, IColInput
-        {
-            public PipelineColumn Input { get; }
-
-            public OutVectorColumn(Vector<TValue> input)
-                : base(Reconciler.Inst, input)
-            {
-                Input = input;
-            }
-        }
-
-        private sealed class OutVarVectorColumn<TValue> : VarVector<bool>, IColInput
-        {
-            public PipelineColumn Input { get; }
-
-            public OutVarVectorColumn(VarVector<TValue> input)
-                : base(Reconciler.Inst, input)
-            {
-                Input = input;
-            }
-        }
-
-        private sealed class Reconciler : EstimatorReconciler
-        {
-            public static Reconciler Inst = new Reconciler();
-
-            private Reconciler() { }
-
-            public override IEstimator<ITransformer> Reconcile(IHostEnvironment env,
-                PipelineColumn[] toOutput,
-                IReadOnlyDictionary<PipelineColumn, string> inputNames,
-                IReadOnlyDictionary<PipelineColumn, string> outputNames,
-                IReadOnlyCollection<string> usedNames)
-            {
-                var columnPairs = new (string input, string output)[toOutput.Length];
-                for (int i = 0; i < toOutput.Length; ++i)
-                {
-                    var col = (IColInput)toOutput[i];
-                    columnPairs[i] = (inputNames[col.Input], outputNames[toOutput[i]]);
-                }
-                return new MissingValueIndicatorEstimator(env, columnPairs);
-            }
-        }
-
-        /// <summary>
-        /// Produces a column of boolean entries indicating whether input column entries were missing.
-        /// </summary>
-        /// <param name="input">The input column.</param>
-        /// <returns>A column indicating whether input column entries were missing.</returns>
-        public static Scalar<bool> IsMissingValue(this Scalar<float> input)
-        {
-            Contracts.CheckValue(input, nameof(input));
-            return new OutScalar<float>(input);
-        }
-
-        /// <summary>
-        /// Produces a column of boolean entries indicating whether input column entries were missing.
-        /// </summary>
-        /// <param name="input">The input column.</param>
-        /// <returns>A column indicating whether input column entries were missing.</returns>
-        public static Scalar<bool> IsMissingValue(this Scalar<double> input)
-        {
-            Contracts.CheckValue(input, nameof(input));
-            return new OutScalar<double>(input);
-        }
-
-        /// <summary>
-        /// Produces a column of boolean entries indicating whether input column entries were missing.
-        /// </summary>
-        /// <param name="input">The input column.</param>
-        /// <returns>A column indicating whether input column entries were missing.</returns>
-        public static Vector<bool> IsMissingValue(this Vector<float> input)
-        {
-            Contracts.CheckValue(input, nameof(input));
-            return new OutVectorColumn<float>(input);
-        }
-
-        /// <summary>
-        /// Produces a column of boolean entries indicating whether input column entries were missing.
-        /// </summary>
-        /// <param name="input">The input column.</param>
-        /// <returns>A column indicating whether input column entries were missing.</returns>
-        public static Vector<bool> IsMissingValue(this Vector<double> input)
-        {
-            Contracts.CheckValue(input, nameof(input));
-            return new OutVectorColumn<double>(input);
-        }
-
-        /// <summary>
-        /// Produces a column of boolean entries indicating whether input column entries were missing.
-        /// </summary>
-        /// <param name="input">The input column.</param>
-        /// <returns>A column indicating whether input column entries were missing.</returns>
-        public static VarVector<bool> IsMissingValue(this VarVector<float> input)
-        {
-            Contracts.CheckValue(input, nameof(input));
-            return new OutVarVectorColumn<float>(input);
-        }
-
-        /// <summary>
-        /// Produces a column of boolean entries indicating whether input column entries were missing.
-        /// </summary>
-        /// <param name="input">The input column.</param>
-        /// <returns>A column indicating whether input column entries were missing.</returns>
-        public static VarVector<bool> IsMissingValue(this VarVector<double> input)
-        {
-            Contracts.CheckValue(input, nameof(input));
-            return new OutVarVectorColumn<double>(input);
-        }
-    }
 }
